change to double

+++ HelpMyStreet.Utils/HelpMyStreet.Contracts/UserService/Request/VolunteerFilter.cs
@@ -9,6 +9,6 @@ namespace HelpMyStreet.Contracts.UserService.Request
     {
         public string Postcode { get; set; }
         public List<SupportActivities> Activities { get; set; }
-        public int? OverrideVolunteerRadius { get; set; }
+        public double? OverrideVolunteerRadius { get; set; }
     }
 }
